Erp-12::Customer Customize the column as per the db changes And CustomerAddress Grid Specific Customer Data Display

+++ MidCapERP.DataAccess/Interface/ICustomerAddressesDA.cs
@@ -6,12 +6,12 @@ namespace MidCapERP.DataAccess.Interface
     {
         public Task<IQueryable<CustomerAddresses>> GetAll(CancellationToken cancellationToken);
 
-        public Task<CustomerAddresses> GetById(int Id, CancellationToken cancellationToken);
+        public Task<CustomerAddresses> GetById(Int64 Id, CancellationToken cancellationToken);
 
         public Task<CustomerAddresses> CreateCustomerAddress(CustomerAddresses model, CancellationToken cancellationToken);
 
-        public Task<CustomerAddresses> UpdateCustomerAddress(int Id, CustomerAddresses model, CancellationToken cancellationToken);
+        public Task<CustomerAddresses> UpdateCustomerAddress(Int64 Id, CustomerAddresses model, CancellationToken cancellationToken);
 
-        public Task<CustomerAddresses> DeleteCustomerAddress(int Id, CancellationToken cancellationToken);
+        public Task<CustomerAddresses> DeleteCustomerAddress(Int64 Id, CancellationToken cancellationToken);
     }
 }
